+++ apps/router-demo/router-remote5-2005/CHANGELOG.md
@@ -1,5 +1,19 @@
 # remote5
 
+## 1.1.26
+
+### Patch Changes
+
+- @module-federation/bridge-react@0.22.1
+- @module-federation/rsbuild-plugin@0.22.1
+
+## 1.1.25
+
+### Patch Changes
+
+- @module-federation/bridge-react@0.22.0
+- @module-federation/rsbuild-plugin@0.22.0
+
 ## 1.1.24
 
 ### Patch Changes