+++ apps/website-new/CHANGELOG.md
@@ -1,5 +1,19 @@
 # website-new
 
+## 1.3.8
+
+### Patch Changes
+
+- @module-federation/error-codes@0.22.1
+- @module-federation/rspress-plugin@0.22.1
+
+## 1.3.7
+
+### Patch Changes
+
+- @module-federation/error-codes@0.22.0
+- @module-federation/rspress-plugin@0.22.0
+
 ## 1.3.6
 
 ### Patch Changes