+++ packages/error-codes/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @module-federation/error-codes
 
+## 0.22.1
+
+## 0.22.0
+
 ## 0.21.6
 
 ## 0.21.5