+++ packages/esbuild/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @module-federation/esbuild
 
+## 0.0.91
+
+### Patch Changes
+
+- @module-federation/sdk@0.22.1
+
+## 0.0.90
+
+### Patch Changes
+
+- @module-federation/sdk@0.22.0
+
 ## 0.0.89
 
 ### Patch Changes