+++ packages/rsbuild-plugin/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @module-federation/rsbuild-plugin
 
+## 0.22.1
+
+### Patch Changes
+
+- @module-federation/enhanced@0.22.1
+- @module-federation/sdk@0.22.1
+- @module-federation/node@2.7.27
+
+## 0.22.0
+
+### Patch Changes
+
+- @module-federation/enhanced@0.22.0
+- @module-federation/sdk@0.22.0
+- @module-federation/node@2.7.26
+
 ## 0.21.6
 
 ### Patch Changes